Some fixups (#12)
* Don't require active connection when probing network devices

This causes the WiFi tile in quick settings to incorrectly report
off. Furthermore, with this check, the same tile is rendered useless.

* Fix the way the icon is fetched in the QS tile when WiFi is off

Otherwise it causes a really long string which in turns causes the
quick settings render outside of the display.

* Truncate SSID in the WiFi tile to 11 characters

It should prints the first 10 characters of the SSID followed by an
ellipsis as the 11th character.

* Fix the feedback loop for brightness

This is what causes the display to flicker and the shell to freeze.

Fixes: #8

* Port over truncation of the WiFi SSID logic to the bar

* Commonize truncation function

* Truncate the text shown in playerctl widget and panel

---------

Co-authored-by: Bahar KURT <kurtbahartr@users.noreply.github.com>